reorganize repository to prepare for kcube implementation
Issue #1
-rename from thorlabs_tcube to thorlabs_cube for generality
-separate and rename base class to _Cube for generality
-separate driver.py into smaller logical modules
-update documentation with new name and package structure
-fix dependencies in setup.py
-add .gitignore
