+++ .gitignore
@@ -1,0 +1,9 @@
+# files
+**/.DS_Store
+
+# file types
+
+# directories
+**/__pycache__/
+doc/_build/
+**/**.egg-info/
